package info javadoc and ValContainer
