--- app/src/main/java/edu/mills/findeatfood/models/ValContainer.java
@@ -1,20 +1,0 @@
-package edu.mills.findeatfood.models;
-
-public class ValContainer<T> {
-    private T val;
-
-    public ValContainer() {
-    }
-
-    public ValContainer(T v) {
-        this.val = v;
-    }
-
-    public T getVal() {
-        return val;
-    }
-
-    public void setVal(T val) {
-        this.val = val;
-    }
-}
