+++ app/src/main/java/edu/mills/findeatfood/models/package-info.java
@@ -1,0 +1,4 @@
+/**
+ * Models for parsing JSON into Java objects.
+ */
+package edu.mills.findeatfood.models;
