+++ app/src/main/java/edu/mills/findeatfood/package-info.java
@@ -1,0 +1,5 @@
+/**
+ * Implementation of the mobile application FINDEATFOOD. Created
+ * for Mills CS 115: Mobile Application Development, Spring 2017.
+ */
+package edu.mills.findeatfood;
